+++ packages/agtree/src/index.ts
@@ -229,6 +229,7 @@ export {
 export {
     Platform,
     PlatformExpressionEvaluator,
+    ProductCode,
     modifiersCompatibilityTable,
     redirectsCompatibilityTable,
     scriptletsCompatibilityTable,